BrainloxChatbot-API/requirements.txt
flask
langchain-community
langchain-google-genai
faiss-cpu
unstructured
google-generativeai